+++ GA_String/Population.cs
@@ -54,7 +54,7 @@ namespace GA_String
         {
             this.matingPool = new List<Chromosome>();
 
-            int bestFitness = 0;
+            double bestFitness = 0;
             for (int i = 0; i < this.popSize; i++) // Finder ud af hvor meget fitness det bedste kromosom har
             {
                 if (this.population[i].fitness > bestFitness)
@@ -66,7 +66,7 @@ namespace GA_String
             // Gennemgår alle kromosomerne i befolkningen, og tilføjer dem til matingPool, baseret på deres fitness-værdi
             for (int i = 0; i < this.popSize; i++)
             {
-                double popFitness = (double)this.population[i].fitness;     // Konverterer fitness-værdien for et kromosom til double
+                double popFitness = this.population[i].fitness;     // Konverterer fitness-værdien for et kromosom til double
                 double matingWeight = popFitness.Map(0, bestFitness, 0, 1); // Mapper fitness-værdien fra 0 til 1, baseret på den bedste fitness-værdi
                 int matingCount = (int)matingWeight * 100;                    // Ganger det med et forholdsvist magisk tal for at det "bliver til noget"
 
@@ -88,19 +88,6 @@ namespace GA_String
 
         public void Reproduce()
         {
-
-            //TODO: Sorter befolkningen, så dele af den kan gøres helt tilfældig for at holde variationen i gang
-
-            //for (int i = 0; i < this.popSize - (this.popSize * 0.3); i++)
-            //{
-                
-            //}
-
-            //for (int i = (int)(this.popSize - (this.popSize * 0.3)); i < this.popSize; i++)
-            //{
-            //    population[i] = new Chromosome(this.target.Length, this.rand);
-            //}
-
             for (int i = 0; i < this.popSize; i++)
             {
                 int parentIndexA = this.rand.Next(this.matingPool.Count); // Forælder A index i matingPool
